Moved some logic for the test suite to a trait.
