+++ tests/Unit/Console/Commands/SyncOrganizationTeamsCommandTest.php
@@ -8,27 +8,12 @@ use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Http;
 use Wotta\SentryTile\Tests\TestCase;
 use Wotta\SentryTile\Tests\Traits\InteractWithFixture;
+use Wotta\SentryTile\Tests\Unit\Traits\InteractWithSentry;
 
 class SyncOrganizationTeamsCommandTest extends TestCase
 {
-    use InteractWithFixture;
-
-    private string $organization;
-
-    /**
-     * All tests tests that should be run here need to have an api token set.
-     * Testing the command with no api token is not needed because the
-     * client test already makes sure the correct error is thrown.
-     */
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->organization = 'test-organization';
-
-        $this->app['config']->set('dashboard.tiles.sentry.token', 'test_api_token');
-        $this->app['config']->set('dashboard.tiles.sentry.base_url', 'https://sentry.io/api/0/');
-    }
+    use InteractWithSentry,
+        InteractWithFixture;
 
     public function test_command_display_information_that_is_stored(): void
     {
